% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/americas-package.R
\docType{package}
\name{americas-package}
\alias{americas-package}
\title{americas: Easy Interactive Maps of United States Survey Data}
\description{
The package exposes one main function, \code{\link[=americas]{americas()}}, which launches an
interactive Shiny dashboard mapping a data frame of longitude/latitude
points over US state boundaries.
}
\author{
\strong{Maintainer}: Ishara Wijayaratne \email{ishara.wijayaratne2512@gmail.com}

}
\keyword{internal}
